update the struct test so that field c of the test struct is of type double

+++ src/test/struct/struct.h
@@ -4,7 +4,7 @@ struct testStruct
 {
     long a;
     bool b;
-    long c;
+    double c;
 };
 
 struct testStructRec
@@ -13,16 +13,16 @@ struct testStructRec
     struct testStructRec *b;
 };
 
-struct testStruct* getTestStruct(long a, bool b, long c);
+struct testStruct* getTestStruct(long a, bool b, double c);
 struct testStructRec* getTestStructRec(long a, struct testStructRec* b);
 
 long getStructFieldA(struct testStruct *);
 bool getStructFieldB(struct testStruct *);
-long getStructFieldC(struct testStruct *);
+double getStructFieldC(struct testStruct *);
 
 struct testStruct *setStructFieldA(struct testStruct *t, long a);
 struct testStruct *setStructFieldB(struct testStruct *t, bool b);
-struct testStruct *setStructFieldC(struct testStruct *t, long c);
+struct testStruct *setStructFieldC(struct testStruct *t, double c);
 
 long getStructFieldRecA(struct testStructRec *t);
 struct testStructRec *getStructFieldRecB(struct testStructRec *t);
